+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+* Add `exchangeRateQuoteId` to `TransactionRequest`
+* Add `exchangeRateQuoteId` to `Transaction`
+
+
 ## 3.10.0
 * Add `dataOnlyRequested` to `ThreeDSecureLookupRequest`
 * Add `paymentReaderCardDetails` to `TransactionCreditCardRequest`